Consolidate GitHub Actions helpers into src/tools/github.ts
Moves GitHub Actions workflow command helpers from tests/ to src/tools/ so they can be used throughout the codebase to reduce ungrouped log lines. The consolidated helpers include proper escaping for workflow commands.
